move createSaltoElementError func to adapter-utils
