+++ packages/adapter-api/src/error.ts
@@ -28,27 +28,4 @@ export const isSaltoElementError = (error: SaltoError | SaltoElementError): erro
 
 export const isSaltoError = (error: unknown): error is SaltoError =>
   _.isObject(error) && 'message' in error && 'severity' in error
-
-export const createSaltoElementErrorFromError = ({
-  error,
-  severity,
-  elemID,
-}: {
-  error: Error
-  severity: SeverityLevel
-  elemID: ElemID
-}): SaltoElementError => ({ message: error.message, detailedMessage: error.message, severity, elemID })
-
-export const createSaltoElementError = ({
-  message,
-  detailedMessage,
-  severity,
-  elemID,
-}: {
-  message: string
-  detailedMessage: string
-  severity: SeverityLevel
-  elemID: ElemID
-}): SaltoElementError => ({ message, detailedMessage, severity, elemID })
-
 export class CredentialError extends Error {}
