+++ packages/adapter-api/test/utils.test.ts
@@ -19,7 +19,6 @@ import {
 } from '../src/elements'
 import { ElemID } from '../src/element_id'
 import { BuiltinTypes } from '../src/builtins'
-import { isSaltoElementError } from '../src/error'
 
 describe('Test utils.ts & isXXX in elements.ts', () => {
   const innerType = new ObjectType({
@@ -232,37 +231,6 @@ describe('Test utils.ts & isXXX in elements.ts', () => {
     })
     it('should return undefined on path to annotation', async () => {
       expect(await getFieldType(mockObjectType, ['someAnno', 'innerField'])).toBeUndefined()
-    })
-  })
-  describe('test error verification', () => {
-    it('should return true when given an error that has elemID', () => {
-      expect(
-        isSaltoElementError({
-          message: '',
-          severity: 'Error',
-          elemID: new ElemID(''),
-          detailedMessage: '',
-        }),
-      ).toBeTruthy()
-    })
-    it('should return false when given an error that has elemID which is undefined', () => {
-      expect(
-        isSaltoElementError({
-          message: '',
-          severity: 'Error',
-          elemID: undefined,
-          detailedMessage: '',
-        }),
-      ).toBeFalsy()
-    })
-    it('should return false when given an error that does not have elemID', () => {
-      expect(
-        isSaltoElementError({
-          message: '',
-          severity: 'Error',
-          detailedMessage: '',
-        }),
-      ).toBeFalsy()
     })
   })
 })
