+++ packages/adapter-utils/index.ts
@@ -23,5 +23,5 @@ export * from './src/elem_id_discrepancy'
 export * from './src/important_values'
 export * from './src/additional_properties'
 export * from './src/config_options'
-export * from './src/error_messages'
+export * from './src/error'
 export * from './src/html_parser'
