+++ packages/zendesk-adapter/src/filters/guide_order/guide_order_utils.ts
@@ -19,8 +19,8 @@ import {
   ListType,
   BuiltinTypes,
   SaltoElementError,
-  createSaltoElementError,
 } from '@salto-io/adapter-api'
+import { createSaltoElementError } from '@salto-io/adapter-utils'
 import { collections, values as lowerDashValues } from '@salto-io/lowerdash'
 import _ from 'lodash'
 import { elements as elementsUtils, fetch as fetchUtils } from '@salto-io/adapter-components'
